app/requirements.txt
flask==1.1.4
itsdangerous==1.1.0
markupsafe==1.1.1
flask-cors
uWSGI==2.0.17.1
Werkzeug==0.16.0
requests
transformers[tf-cpu]
mysql-connector-python